requirements.txt
flask==2.3.3
requests==2.31.0
requests-oauthlib==1.3.1
oauthlib==3.2.2
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
gunicorn==21.2.0
beautifulsoup4==4.12.2
sqlalchemy==2.0.23
lxml==4.9.3